Modify viewport size, color, profile image & cards
Modified viewport size via grid spacing as the prior grid template
left a blank white "footer" at the bottom of the page, with no actual
footer HTML. Changed the background color of the sidebar to match
the assignment example. Changed the background color of the main
section to match the color of the search bar, although it still doesn't
match the color on the assignment example. Added box shadows to all the
cards in the main section. Modified font styling in the Trending card
section.